requirements.txt
annotated-types==0.7.0
anyio==4.6.0
appdirs==1.4.4
asgiref==3.4.1
async-generator==1.10
attrs==22.1.0
Babel==2.9.1
backports.entry-points-selectable==1.1.1
backports.tarfile==1.2.0
bcrypt==4.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.7
click==8.1.7
colorama==0.4.6
cryptography==3.4.8
cycler==0.11.0
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.9
Django==2.0
dnspython==2.7.0
docopt==0.6.2
docutils==0.16
ebaysdk==2.1.5
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.0
filelock==3.16.1
fonttools==4.28.1
freezegun==0.3.15
gevent==21.8.0
git-filter-repo==2.45.0
greenlet==1.1.2
h11==0.13.0
hatchling==1.25.0
httpcore==1.0.6
httpx==0.27.2
hyperlink==21.0.0
idna==2.10
importlib_metadata==8.5.0
iniconfig==2.0.0
isodate==0.6.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==2.11.3
keyring==25.4.1
kiwisolver==1.3.2
libsass==0.20.1
lxml==4.6.5
markdown-it-py==3.0.0
MarkupSafe==1.1.1
matplotlib==3.5.0
mdurl==0.1.2
more-itertools==10.5.0
num2words==0.5.9
numpy==1.21.4
ofxparse==0.21
outcome==1.2.0
packaging==24.1
pandas==1.3.4
passlib==1.7.4
pathspec==0.12.1
pexpect==4.9.0
Pillow==9.0.1
platformdirs==4.3.6
plotly==5.4.0
pluggy==1.5.0
polib==1.1.0
psutil==5.8.0
psycopg2==2.9.5
ptyprocess==0.7.0
pyasn1==0.6.1
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
pydot==1.4.2
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.10.1
pyOpenSSL==20.0.1
pyparsing==3.0.6
PyPDF2==1.26.0
pypiwin32==223
pyserial==3.5
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.8.1
python-dotenv==1.0.1
python-geofs==0.1.5
python-jose==3.3.0
python-multipart==0.0.12
python-stdnum==1.16
pytz==2021.3
pyusb==1.0.2
pywin32==306
pywin32-ctypes==0.2.3
qrcode==6.1
reportlab==3.5.59
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==1.0.0
rich==13.9.2
rsa==4.9
selenium==4.4.0
setuptools-scm==6.3.2
shellingham==1.5.4
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
starlette==0.38.6
tenacity==8.0.1
tomli==1.2.2
tomli_w==1.1.0
tomlkit==0.13.2
trio==0.21.0
trio-websocket==0.9.2
trove-classifiers==2024.9.12
typing_extensions==4.12.2
urllib3==1.26.5
userpath==1.9.2
uv==0.4.20
uvicorn==0.31.1
virtualenv==20.26.6
vobject==0.9.6.1
Werkzeug==2.0.2
wsproto==1.1.0
xlrd==1.2.0
XlsxWriter==1.1.2
xlwt==1.3.0
zeep==4.0.0
zipp==3.20.2
zope.event==5.0
zope.interface==6.0
zstandard==0.23.0
